requirements.txt
# ~~ Generated by projen. To modify, edit .projenrc.py and run "npx projen".
fastapi
google-cloud-storage
redis
uvicorn[standard]

opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-gcp-trace
opentelemetry-instrumentation-requests
# de
